fix ci

+++ app/src/items.c
@@ -273,7 +273,7 @@ static items_error_t items_storeAllTransfers() {
 
                 array_get_element_count(json_all, clist_token_index, &clist_element_count);
 
-                for (uint8_t i = 0; i < clist_element_count; i++) {
+                for (uint16_t i = 0; i < clist_element_count; i++) {
                     if (array_get_nth_element(json_all, clist_token_index, i, &token_index) == parser_ok) {
                         uint16_t name_token_index = 0;
                         if (object_get_value(json_all, token_index, JSON_NAME, &name_token_index) == parser_ok) {
